move Attendee to Response package + remove useless test
